requirements.txt
google-generativeai
google-genai
python-dotenv
rich
readchar
google-auth-oauthlib
google-auth-httplib2
duckduckgo-search

